refactor: 폴더 및 파일 구조 네이밍 수정 - 폴더 : 모든 폴더는 kebab case로 명명한다. - 파일 : 컴포넌트 파일(tsx)은 pascal case, 함수 파일(ts)는 camel case로 명명한다.
