+++ src/entities/README.md
@@ -19,6 +19,11 @@ entities/
     └── index.ts      # 외부로 노출할 모듈을 export (Public API)
 ```
 
+## 폴더 및 파일명 규칙
+
+- **폴더** : 폴더 이름은 kebab case로 작성한다.
+- **파일** : 파일 이름은 .tsx 와 같은 컴포넌트 단위인 경우 Pascal case로 작성하고, .ts와 같은 기능 단위는 camel case로 작성한다.
+
 ## ✍️ 작성 규칙
 
 - **독립성**: `entities`는 오직 `shared` 레이어에만 의존할 수 있습니다. `features`나 `widgets`를 import해서는 안 됩니다.
