+++ src/entities/tanstack-test/model/hooks/useTanstackQuery.ts
@@ -1,5 +1,5 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { getTanstack } from '@/entities/TanstackTest/api/getTanstack';
+import { getTanstack } from '@/entities/tanstack-test/api/getTanstack';
 import TanstackInfo from '../types';
 
 export function useTanstackQuery(): UseQueryResult<TanstackInfo | null, Error> {
