+++ src/entities/tasting-list/model/hooks/useTastingListQuery.ts
@@ -1,5 +1,5 @@
 import { useQuery, UseQueryResult } from '@tanstack/react-query';
-import { getTastingList } from '@/entities/TastingList/api/getTastingList';
+import { getTastingList } from '@/entities/tasting-list/api/getTastingList';
 import { TastingList } from '../types';
 
 export function useTastingListQuery(meetingId: number): UseQueryResult<TastingList | null, Error> {
