+++ src/features/axios-test/api/deleteTest.ts
@@ -1,5 +1,5 @@
-import { TestPost } from '@/features/AxiosTest/model/types';
 import { deleteApi } from '@/shared/api/axiosApis';
+import { TestPost } from '@/features/axios-test/model/types';
 
 export async function deleteTest(): Promise<TestPost | null> {
   return deleteApi<TestPost>('/posts/1', 'https://jsonplaceholder.typicode.com/');
