+++ src/features/axios-test/model/hooks/useDeleteTestMutation.ts
