+++ src/shared/README.md
@@ -4,8 +4,8 @@
 
 ## 🎯 핵심 역할
 
-* 애플리케이션의 특정 도메인과 완전히 분리된, 재사용 가능한 코드를 제공합니다.
-* UI Kit, 공통 라이브러리 설정, 헬퍼 함수 등을 관리합니다.
+- 애플리케이션의 특정 도메인과 완전히 분리된, 재사용 가능한 코드를 제공합니다.
+- UI Kit, 공통 라이브러리 설정, 헬퍼 함수 등을 관리합니다.
 
 ## 📁 폴더 구조
 
@@ -18,12 +18,17 @@ shared/
 └── index.ts      # 각 모듈을 export
 ```
 
+## 폴더 및 파일명 규칙
+
+- **폴더** : 폴더 이름은 kebab case로 작성한다.
+- **파일** : 파일 이름은 .tsx 와 같은 컴포넌트 단위인 경우 Pascal case로 작성하고, .ts와 같은 기능 단위는 camel case로 작성한다.
+
 ## ✍️ 작성 규칙
 
-* **독립성**: `shared` 레이어는 프로젝트 내 어떤 다른 레이어에도 의존해서는 안 됩니다. **의존성이 없어야 합니다.**
-* **일반성**: 여기에 작성되는 코드는 이 프로젝트가 아닌 다른 프로젝트에 가져다 써도 문제가 없을 정도로 일반적이어야 합니다.
-* **Public API**: `shared/ui/index.ts` 처럼 각 하위 폴더마다 `index.ts`를 두어 외부로 노출할 컴포넌트나 함수를 명시적으로 관리하는 것이 좋습니다.
+- **독립성**: `shared` 레이어는 프로젝트 내 어떤 다른 레이어에도 의존해서는 안 됩니다. **의존성이 없어야 합니다.**
+- **일반성**: 여기에 작성되는 코드는 이 프로젝트가 아닌 다른 프로젝트에 가져다 써도 문제가 없을 정도로 일반적이어야 합니다.
+- **Public API**: `shared/ui/index.ts` 처럼 각 하위 폴더마다 `index.ts`를 두어 외부로 노출할 컴포넌트나 함수를 명시적으로 관리하는 것이 좋습니다.
 
 ## ❌ 금지 사항
 
-* 특정 비즈니스 로직이나 도메인 지식을 담고 있는 코드를 작성하지 마세요. (e.g., `Post` 타입을 직접 참조하거나, '게시글을 불러오는' 훅을 만들면 안 됩니다.)
+- 특정 비즈니스 로직이나 도메인 지식을 담고 있는 코드를 작성하지 마세요. (e.g., `Post` 타입을 직접 참조하거나, '게시글을 불러오는' 훅을 만들면 안 됩니다.)
